Avoid exclusive language
Many years ago I wrote on a whiteboard something like

    blacklist = bad
    whitelist = good

and then to save space, rubbed out "list".
And I went "uh-oh, this looks super fucking shady if read out-of-context".

This codebase predates that epiphany, and
I hadn't noticed it was still problematic.

Back then I considered "allow" and "deny", which
is what the rest of the industry has since standardized on.
On fixed-width terminals, those strings don't align neatly.
I still interact primarily with fixed-width terminals.
So I settled on "shit" and "good", and still use those.
https://en.wiktionary.org/wiki/shitlist

                                * * *

https://en.wikipedia.org/wiki/Inclusive_language
https://en.wiktionary.org/wiki/blacklist#Usage_notes
https://en.wikipedia.org/wiki/Blacklisting#Origins_of_the_term

There is no etymological connection between "blacklist" and "black people",
but continuing to use "blacklist" is like continuing to use
"niggardly" when "miserly" exists.

This change does fuck-all to reduce systemic inequality.
It's popular with corporate America because it's cheap PR.
But it's so cheap that it's easier to do it than argue about it.
And it probably still a net benefit. [FIXME: find some peer-reviewed evidence.]

If you want to actually improve equality globally, you could try

 * Pass Doha so that agricultural goods get
   the same protections as manufactured goods.
   The current setup gives developed countries a built-in advantage.
   https://en.wikipedia.org/wiki/Doha_Development_Round#Agriculture

 * Pass UBI so people aren't forced to choose between dignity and survival.
   https://en.wikipedia.org/wiki/Universal_basic_income

 * Pass death tax and other progressive taxes,
   to mitigate intergenerational wealth concentration.
   https://en.wikipedia.org/wiki/Estate_tax

 * Think of some way to reduce agro slavery, e.g.
   https://foodispower.org/human-labor-slavery/slavery-chocolate/
   Raising the cost of enslaving humans,
   lowering the cost of machines,
   improving the adaptability of machines
   (manufactured goods are all identical, grown plants aren't).

 * Education.  [FIXME: more details here.]
   https://ncase.me/polygons/

 * Hang out with people who aren't like you.

https://en.wikipedia.org/wiki/Scientific_racism
https://en.wikipedia.org/wiki/Structural_inequality
https://en.wikipedia.org/wiki/Economic_inequality
https://en.wikipedia.org/wiki/Social_inequality
https://en.wikipedia.org/wiki/Institutional_racism

"Help that woman. Help that man. That's what brothers are for, brother."
https://genius.com/Gil-scott-heron-brother-annotated
https://genius.com/Gil-scott-heron-evolution-and-flashback-annotated

"I thought by now all of this should have been taken care of long ago.
When I was a boy I didn't think I'd be a grown man looking at it still."
https://genius.com/Mavis-staples-my-own-eyes-lyrics

PS: I left the string "whitespace" as-is. >_>